change 'customer_id' data type to 'BIGSERIAL' instead

+++ src/main/resources/schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS customers (
-    customer_id SERIAL PRIMARY KEY NOT NULL,
+    customer_id BIGSERIAL PRIMARY KEY NOT NULL,
     name VARCHAR(100) NOT NULL,
     email VARCHAR(100) NOT NULL,
     mobile_number VARCHAR(20) NOT NULL,
@@ -10,8 +10,8 @@ CREATE TABLE IF NOT EXISTS customers (
 );
 
 CREATE TABLE IF NOT EXISTS accounts (
-    account_number SERIAL PRIMARY KEY NOT NULL,
-    customer_id INT NOT NULL,
+    account_number BIGSERIAL PRIMARY KEY NOT NULL,
+    customer_id INT8 NOT NULL,
     account_type VARCHAR(100) NOT NULL,
     branch_address VARCHAR(200) NOT NULL,
     created_at DATE NOT NULL,
